Final Project/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.0
aiosignal==1.4.0
attrs==25.4.0
beautifulsoup4==4.14.2
bleach==6.3.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.3
contourpy==1.3.3
curl_cffi==0.13.0
cycler==0.12.1
dateparser==1.2.2
filelock==3.20.0
fonttools==4.60.0
fredapi==0.5.2
frozendict==2.4.6
frozenlist==1.8.0
fsspec==2025.10.0
hf-xet==1.2.0
huggingface-hub==0.36.0
idna==3.10
joblib==1.5.2
kaggle==1.7.4.5
kiwisolver==1.4.9
lxml==6.0.2
matplotlib==3.10.6
multidict==6.7.0
multitasking==0.0.12
numpy==2.3.3
packaging==25.0
pandas==2.3.2
patsy==1.0.1
peewee==3.18.2
pillow==11.3.0
platformdirs==4.4.0
praw==7.8.1
prawcore==2.4.0
propcache==0.4.0
protobuf==6.32.1
pycparser==2.23
pycryptodome==3.23.0
pyparsing==3.2.5
python-binance==1.0.29
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytz==2025.2
PyYAML==6.0.3
regex==2025.9.18
requests==2.32.5
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.16.2
seaborn==0.13.2
setuptools==80.9.0
six==1.17.0
soupsieve==2.8
statsmodels==0.14.5
text-unidecode==1.3
threadpoolctl==3.6.0
tokenizers==0.22.1
tqdm==4.67.1
transformers==4.57.3
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
update-checker==0.18.0
urllib3==2.5.0
vaderSentiment==3.3.2
webencodings==0.5.1
websocket-client==1.9.0
websockets==15.0.1
yarl==1.22.0
yfinance==0.2.66
